python/requirements.txt
blinker==1.7.0
cffi==1.16.0
click==7.1.2
ffmpeg==1.4
ffprobe==0.5
Flask==1.1.4
Flask-Sockets==0.2.1
gevent==23.9.1
gevent-websocket==0.10.1
greenlet==3.0.2
gunicorn==21.2.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==2.0.1
numpy==1.26.2
packaging==23.2
pycparser==2.21
pydub==0.25.1
six==1.16.0
Werkzeug==1.0.1
zope.event==5.0
zope.interface==6.1
pyyaml==6.0.1